Refactor more functions and components out of EnrollmentIndex, start refactoring UserListUI.
