+++ app/frontend/components/AtCapacity.js
@@ -1,5 +1,5 @@
 function AtCapacity(limit, number) {
   return limit <= number;
-}
+};
 
 export default AtCapacity;
