+++ app/frontend/components/EnrollmentIndex.jsx
@@ -1,9 +1,15 @@
 import React, { useState } from 'react';
 import "bootstrap/dist/css/bootstrap.min.css";
 import AtCapacity from '~/components/AtCapacity.js'
-import { Button, Card, Col, Row, Alert } from "react-bootstrap";
+import { Button, Card, Col, Row} from "react-bootstrap";
 import DeleteConfirmation from '~/components/DeleteConfirmation.jsx';
 import DisplayDeleteSuccess from '~/components/DisplayDeleteSuccess.jsx';
+import NewStudentError from '~/components/NewStudentError.jsx';
+import endTime from '~/components/EndTime.js';
+import LinkRendering from '~/components/LinkRendering.js';
+import StudentText from '~/components/StudentText.jsx';
+import UserHeaderUI from '~/components/UserHeaderUI.jsx';
+// import UserListUI from '~/components/UserListUI.jsx';
 
 function EnrollmentIndex(props) {
   const csrf_token = document.head.getElementsByTagName('meta')[2].content;
@@ -11,16 +17,10 @@ function EnrollmentIndex(props) {
   const sheridanEnrollments = JSON.parse(document.getElementById("data").getAttribute("sheridan_enrollments"));
   const smokyEnrollments = JSON.parse(document.getElementById("data").getAttribute("smoky_enrollments"));
   const newStudentError = document.getElementById("data").getAttribute("new_student_error");
+  const deleteConfirmation = document.getElementById("data").getAttribute("delete_confirmation");
   const [displayConfirmationModal, setDisplayConfirmationModal] = useState(false);
   const [deleteMessage, setDeleteMessage] = useState(null);
-  const deleteConfirmation = document.getElementById("data").getAttribute("delete_confirmation");
   const [deletePath, setDeletePath] = useState(null);
-
-  function ShowDeleteModal(enrollment, id) {
-    setDeletePath(`/admin/enrollments/${id}`);
-    setDeleteMessage(`Are you sure you want to delete the enrollment in ${enrollment.location} on ${enrollment.date} at ${enrollment.time}`);
-    setDisplayConfirmationModal(true);
-  }
 
   const hideConfirmationModal = () => {
     setDisplayConfirmationModal(false);
@@ -30,33 +30,11 @@ function EnrollmentIndex(props) {
     setDisplayConfirmationModal(false);
   };
 
-  function NewStudentError() {
-    if (newStudentError === 'true') {
-      return <Alert variant="danger" dismissible>"This session is no longer available."</Alert>;
-    }
-  }
-
-  function LinkRendering(capacity, id) {
-    if (capacity) {
-      return "/enrollments";
-    } else {
-      return `/enrollments/${id}/students/new`;
-    }
-  }
-
-  function StudentText(capacity, enrollment) {
-    if (capacity) {
-      return <Card.Text>Session is Full</Card.Text>;
-    } else {
-      return <Card.Text>Available seats: {enrollment.student_limit - enrollment.students}</Card.Text>;
-    }
-  }
-
-  function UserHeaderUI(userType) {
-    if (userType === "admin") {
-      return <Card.Header style={{ textAlign: "center" }}><a href={"/admin/enrollments"}>Enrollment Index</a> &emsp; <a href={"/admin/enrollments/new"}>New Enrollment Session</a></Card.Header>;
-    }
-  }
+  function ShowDeleteModal(enrollment, id) {
+    setDeletePath(`/admin/enrollments/${id}`);
+    setDeleteMessage(`Are you sure you want to delete the enrollment in ${enrollment.location} on ${enrollment.date} at ${enrollment.time}`);
+    setDisplayConfirmationModal(true);
+  };
 
   function UserListUI(userType, enrollment) {
     if (userType === "admin") {
@@ -65,9 +43,9 @@ function EnrollmentIndex(props) {
           <Card.Text>Student Limit: {enrollment.student_limit}</Card.Text>
           <Card.Text>Number of Students: {enrollment.students}</Card.Text>
           <Row xs={2} md={2}>
-            <Button disabled={AtCapacity(enrollment.student_limit, enrollment.students)} href={`admin/enrollments/${enrollment.id}/students/new`} size="sm" variant="outline-dark">Add Student</Button>
-            <Button href={`admin/enrollments/${enrollment.id}/edit`} size="sm" variant="outline-dark">Edit Session</Button>
-            <Button href={`admin/enrollments/${enrollment.id}`} size="sm" variant="outline-dark">Show Details</Button>
+            <Button disabled={AtCapacity(enrollment.student_limit, enrollment.students)} href={`/admin/enrollments/${enrollment.id}/students/new`} size="sm" variant="outline-dark">Add Student</Button>
+            <Button href={`/admin/enrollments/${enrollment.id}/edit`} size="sm" variant="outline-dark">Edit Session</Button>
+            <Button href={`/admin/enrollments/${enrollment.id}`} size="sm" variant="outline-dark">Show Details</Button>
             <Button size="sm" variant="outline-dark" onClick={() => ShowDeleteModal(enrollment, enrollment.id)}>Delete Session</Button>
           </Row>
         </div>
@@ -82,26 +60,7 @@ function EnrollmentIndex(props) {
         </div>
       );
     }
-  }
-
-  const endTime = (time) => {
-    let endHour = parseInt(time.slice(0,2)) + 3;
-    const meridiemSwitch = {
-      "P": "A",
-      "A": "P"
-    };
-    if(endHour > 15) {
-      endHour -= 12;
-      return `0${endHour}${time.slice(-6)}`;
-    } else if(endHour > 12 && endHour <= 15) {
-      endHour -= 12;
-      return (('0'+`${endHour}`).slice(-2)+`${time.slice(2,6)}`+`${meridiemSwitch[time[6]]}M`);
-    } else if(endHour === 12) {
-      return (('0'+`${endHour}`).slice(-2)+`${time.slice(2,6)}`+`${meridiemSwitch[time[6]]}M`);
-    } else { 
-      return `${('0'+ (parseInt(time.slice(0,2))+3)).slice(-2)}${time.slice(-6)}`;
-    }
-  }
+  };
 
   const listMayEnrollments = mayEnrollments.map((enrollment) =>
     <Col key={enrollment.id}>
@@ -145,7 +104,7 @@ function EnrollmentIndex(props) {
   return (
     <Card border="light">
       {DisplayDeleteSuccess(deleteConfirmation)}
-      {NewStudentError()}
+      {NewStudentError(newStudentError)}
       {UserHeaderUI(props.admin)}
       <Card.Title style={{ textAlign: "center", margin: "2%" }}>Eloise May Enrollment Sessions:</Card.Title>
       <Row xs={2} md={3} className="g-4 justify-content-center">
