+++ app/frontend/components/LinkRendering.js
@@ -1,0 +1,9 @@
+function LinkRendering(capacity, id) {
+  if (capacity) {
+    return "/enrollments";
+  } else {
+    return `/enrollments/${id}/students/new`;
+  }
+};
+
+export default LinkRendering;
